Purpose: new feature and some bug fixes
Description:
Add a new feature to "skip" some tests.
Sometimes, a machine has problem (e.g. titan and sleipnir having AFS
problems).  Sometimes some feature in a machine is temporary off (e.g.,
PGI compilers in eirene.)  The tests will be reported being SKIPPED
which will be reminder to remove the skipping once the problem is fixed.

I installed the skipping control in snaptest.cfg rather than in
allhostsfile to have a finer control to skip just a subset of the
tests of a machine, rather than skipping the all tests of a machine.

Bug fixes: fixed the miscounting of number of tests ran.  It was
reporting one too many.

Platforms tested:
Only tested in eirene since eirene is the one uses it directly.

Misc. update: