Purpose: Bug fix.
Description:
n_test must be inited to 1 when it is for tests within one machine.
Undid my previous mistake.

Platforms tested:
Eirene.
Did not do the h5committest because this is a script and if it works
in one machine, it most like will work in all Unix platform.

Misc. update: