Purpose: Feature change.
Description:
--prefix default was /usr/local.  This is not approprirate for
development version since it is not likely that one would want
to install a development version in the /usr/local area.

Solution:
Change --prefix default to `pwd`/hdf5, a directory in the build
directory.  This eliminates the problem that the builder may not
have permission to install in the default install directory.
It also allows multiple builds in the same host or multiple srcdir
using the same source.

Platforms tested:
h5committested.
Also tested in eirene by trying various configure combination to
verify the new effect.

Misc. update: