Purpose: Bug fix
Description:
It uses the value of $ARCH as a gcc option but the linux clusters
at NCSA define $ARCH as environment variable with values that are
not a valid compiler option.  That caused the configure to fail
because it was not able to compile at all.

Solution:
Change ARCH to lower case $arch (convention dictates environment
variables are upper cases.)  Also preset $arch to null and do not
honor any pass it values.

Platforms tested:
Attempted to run h5committest but sol was failing due to /tmp
filled.  Copper failed in parallel test but that could be a
system problem since it was totally loaded.  Verbena failed in
h5diff which has nothing to do with the change.
Tested in TG_NCSA (gcc and ecc)--both failed in h5diff.  Again,
nothing to do with the change.  So, it is safe to commit this
change.

Misc. update: