Purpose: bug fixes
Description:

the return error code for a  function was not initialized.
in HP-UX  it happened that this variable was initialized to -1
causing the function to return with an error condtion
solution : initialized the variable to 0

the name of the dataset was printed after the differences in verbose mode
and report when differences were found
solution : check first if differences were found and then
print the name of dataset and differences
in verbose mode always print the name first


Solution:

Platforms tested:
linux
HP-UX
solaris
AIX

Misc. update: