Purpose: Bug fix.
Description:
The patch for running XML test script had undefined variables that
pmake (IRIX64 native make) would not tolerated.  It is bad to duplicate
codes too.

Solution:
Replaced all the duplicated code with recursive call of make with
a redefined TEST_SCRIPTS.  This has the problem of not calling
the XML tests by default but the XML tests failed in some platforms
including the IRIX64.  This solution has the same effect as before
but cleaner.  Long term fix should be done in h5dump or the XML test
scripts.

Platforms tested:
H5committested.  Also in modi4 using pmake.

Misc. update: