Purpose: Update hdf5build.bat file.
Description:
Previouly, when building HDF5 from command line, hdf5 and hdf5dll projects were built three times to build HDF5 C/C++/Fortran libraries.  There was due that the "Rebuild" option was used with msdev.exe.  It is kind of wasting to build hdf5 and hdf5dll projects three time.

Solution:
The batch file is reorganized.  So hdf5 and hdf5dll projects will be only be built once.
Also, the syntax detecting feature is added to the batch file.  Now if an unexpected option was given to the batch file, a warning message and the help messages would be shown.

Platforms tested:
Visual C++ 6.0 on Windows 2000 and XP.

Misc. update: