Purpose: Update.
Description:
Update HDF5 on CodeWarrior project hdf5/proj/codewarrior/hdf5.mcp.

Solution:
1. Use global source trees to specify the root directories for Zlib and Szip external libraries.

Previously, absoulte paths were used to specify where to find Zlib and Szip header and library files in the access path for each and every target in the project.  Using this method, we need to modify the absolute paths for all related targets if theexternal libraries paths are changed, which is not convenient.

Now, the global source threes are used to specify the root directories for Zlib and Szip libraries.  So whenever Zlib and Szip root directories are changed (for example, transfer the project from one machine to another), we just need to modify the definition of the project's source trees to specify the external library paths.

Note: hdf5/release_docs/INSTALL_codewarrior.txt will be updated to instruct how to modify the global source threes.

2. Add libtest target to generate libtest.lib under hdf5/proj/codewarrior/lib/ directory.
3. Add toolslib target to generate toolslib.lib under hdf5/proj/codewarrior/lib/ directory.
4. Add the following library testing targets:
      dangle, filename, reserved, and talign.
5. Add the following tool testing targets:
      h5import, h5importtst, h5jam, j5jamtst, h5ls, h5repart, and h5unjam.
Note: There is a CodeWarrior-specified bug in h5import.c, I am still investigating it with Pedro.  It will be updated once it is fixed.  So now, h5import test is temperarily unavailable on CodeWarrior.

Platforms tested:
CodeWarrior 8.0 on Windows 2000/XP.

Misc. update: