Purpose: Bug fixed.
Description:
Heping used to have a network limit that if too many rsh connections
occurred in a second, it stopped accepting connections for 30 seconds.
That caused testph5diff to fail when too many mpi jobs are executed
repidly.  A sleep 2 was in to go around this problem.  Heping's limit
has been raised, thus no more need for this hack.  Sleep 2 is removed.

Platforms tested:
Tested in heping (pp).

Misc. update: