Purpose: Code cleanup
Description:
On Cray X1, a loop in the Fortran C source confuses the compiler's optimizer
and hangs the build.

Solution:
Documented the fix in the code, and added an #ifdef so that
the optimization is only disabled on Cray X1.

Platforms tested:
mir (fix was already checked into code; this was a minor change)