Purpose: Feature
Description:
Added new function, getenv_all (only for parallel builds)

Solution:
getenv_all is a collective version of getenv.  It is used to the return
the valu e of an environment variable in another task and can be used
to synchronize all of th e environment variables used by all of the
tasks.  This helps to fix problems that commonly crop up (like with
HDF5_PARAPREFIX) when an environment variable does not propogate to o
ther tasks.

Platforms tested:
heping (serial and pp).