Purpose: bug fix.
Description:
In Janus (Tflops), the test would seg-fault in production (compiled
with -O).  If compiled with -O0, it would be okay.  Track down the
problem to the for loop index i & j being declared with type hsize_t
which is usually unsigned long int.  Guess the Tflop compiler has
a bug in optimization that could not handle unsigned int as for loop
index.

Solution:
Since it is no need for i & j to be hsize_t and it makes more
sense to have them with type int.  Changed them to int type.

Platforms tested:
Passed in tflops and colonelk here.