Purpose: Bug fix
Description:
On MCR, the mpich compiler is called mpiicc.  The configure script's check
for whether or not complicated derived datatypes work was based upon
a compiler name of mpicc.  This result in the configure script incorrectly
assuming that complciated derived datatypes worked when they really didn't and caused
the parallel tests to fail.

Solution:
Changed the config file to also check fi the compiler is named mpiicc.

Platforms tested:
MCR

Misc. update: