Purpose: Feature
Description:
Added blurb about future todo's for this test.

Added support for HDF5_PARAPREFIX to determine the directory where the test file
is stored.

Solution:
Used getenv_all to get the value of HDF5_PARAPREFIX.  Note that, if a command-line
parameter is passed to the program to specify a path, it will override the value of
HDF5_PARAPREFIX.

Platforms tested:
copper, colonelk

Misc. update: