Purpose: Minor bug fixes.
Description:
1.  Changed free() calls to HDfree()
2.  Corrected behavior of -m command-line parameter
3.  Changed return value to always return 0.

Solution:

2.  The -m flag tells the test to run only the MPI IO tests.  However, it would incorrectly make
the test run both the MPI and POSIX tests (same as the default behavior).

3.  This test is known to fail on many platforms, and, even on those platforms where it usually passes,
it is known to suffer transient failures (especially with small test file sizes).  It's outcome is
also very dependent on the filesystem on which the testfile is created.  Corrected
the program to always return success, so that it doesn't interfere with the daily tests.

The motivation for this is that
this test has nothing to do with the HDF library and is an auxiliary test.  Failures in t_posix_complaint
do not necessarily mean that parallel HDF will fail, but simply indicate something to look into,
especially on new platforms.  This is now an "output only" test, and any errors will be only be visible
in the output.

Platforms tested:
copper (all of these were minor changes)

Misc. update: