Purpose: Bug fix.
Description:
getenv_all, like some implementations of getenv, keeps an internal buffer
for the getenv values.  Application must make a copy of the content of the
returned pointer before calling getenv_all again IF it wants to use the
first content later.  The previous code violated this rule and caused
error.

Solution:
Did a small revamp of the h5_fixname to fix the above problem by moving
the getenv_all for HDF5_NOCLEANUP eariler to avoid it clashing the
call of getenv_all for HDF5_PARAPREFIX.  No a good permenant fix.  Need
to really recode this routine.

Platforms tested:
h5committested.