Purpose: Bug fix
Description:
Trying to use h5c++ to link against the high-level library failed.

Solution:
Firstly, the libraries were listed in the wrong order on the linking
line in h5c++.in.
Secondly, in the 1.6 branch the C++ interface has its own configure
script which didn't know whether the high-level library was enabled
or disabled, so when h5c++ tests whether the HL interface is being built,
the test always fails.
Both of these problems were fixed (by reordering the libraries and
adding a test for --enable-hl to c++/configure.in).

Platforms tested:
mir, shanti, copper