Purpose: Bug fix.
Description:
C++ testhdf5 failed when it tried to inspect the size of a file
via the HDstat call.  The reason was that the CFLAGS has -D_LARGE_FILES
hardset in it while the CXXFLAGS does not.  So, C is using an off_t
that is 8 bytes while C++ is using off_t as 4 bytes.  C is using
a version of 64bits stat/off_t/etc while C++ uses the 32 bits version.
Thus a failure.

Solution:
Added -D_LARGE_FILES to $CXXFLAGES so that both C and C++ are using
the same version of stat/off_t/etc.

Platforms tested:
Tested in Copper only, both 32 and 64 bits since the changes affected
the AIX platform only.