Multi file driver test fails.
Skipped until fix is found.

Tested on heping with pgf90.

Note:  Logic was wrong; as a result, the test was bypassed for all 
compilers except those (like Intel) that understand !DEC compiler directive. 

I confirmed that the test fails on heping with pgf90. The symptoms are the same:
1.6.5 h5dump can read multi files created by fortran test, 
but built h5dump fails with internal error.

 --This line, and those below, will be ignored--

M    test/fortranlib_test.f90