Purpose ------- Bug Fix.
Description
-----------
When compiling h5repack with icc, it presents the error: "no rule to make target 
'h5trav.h', needed by 'h5repack.lo'. Stop."


Solution
--------
There doesn't appear to be an intuitive reason why it's looking for a rule for 
h5trav.h, so I've just acquiesced to the error's demands and created a rule. 
The rule is blank, so it does nothing, but it now compiles without complaint.


Tested
------
kagiso, smirom