H5FDwindows.h is included in the hdf5.h but we didn't put it into Makefile.am un der /src. So when gmake check-install is run for HDF5 examples, it will fail. Th e fix is to include a macro #ifdef H5_HAVE_WINDOW around #include H5FDwindows.h so that H5FDwindows.h will only be included on windows, not on other platforms.
We've fixed with the same way for other source code that includes H5FDwindows.h.

This has been tested with h5committest. However, it takes a few hours to test all tests and except kagiso. Tests on linew and smirom cannot finish. It takes forever for the fheap test. 

And for parallel test at kagiso, ph5diff test seems to be hanged. I think all of our changes will only affect windows. So in order not to block other developers to check in their code, we just go ahead to check our changes. 

Kent Yang