Purpose: Fix subtle bug in Windows test script
Description:
Our Windows test script was using "ERRORLEVEL" to check the diff output for our tooltests.  However, FC does not actually set ERRORLEVEL with valid input.  As a work around, we pipe the output into FIND, which sets the ERRORLEVEL if it doesn't find a valid string.  We probably never saw this bug before because h5dump sets the ERRORLEVEL.

Tested:
WinXP.