Purpose: Fix a subtle Windows bug in h5unjam
Description:
In h5unjam.c, there was a call to "open", rather than "HDopen".  This isn't a problem on most platforms, as the are equivalent.  However, on Windows, we use the HDopen macro to unconditionally set the O_BINARY flag.  This is needed to safely seek within a file.

Tested:
VS2005 on WinXP
h5committest (kagiso linew smirom)