Purpose: Fix reference in the Windows batch-build script
Description:
In the Windows command line build script, there was a reference to the "typegen" project, which generates h5detect before HDF5 is built.  Now, we integrate h5detect into the overall build process, and typegen has been removed.  This checkin removes the reference in the batch script.  This fixes bugzilla bug 1021.