Bug fix: Up (an IBM SP AIX machine) was failing the following test. Testing BINARY I8 CR LF EOF                                           *FAILED*
Fix:
Pedro deduced it was because of AIX interprets a native char as
unsigned while some other machines interprets it signed and the
expected output to be so. Pedro sent me a fix for this problem
and it worked for Up.  Per Pedro's instruction, I am committing
the fix directly from the machine zeus. (up does not have svn.
Zeus, another machine at this site, supports svn.)

Tested:
Up.