Purpose: Add hdf5 library as a dependency to hdf5_hl library in Windows project files
Description:
For some reason, the hdf5_hl library didn't have hdf5 listed as a dependency.  This isn't normally a problem, because building "all" will build hdf5 library.  However, when a user tries to build only hdf5_hl, hdf5 library wouldn't be built.

Although not technically required for compiling the library code, the hdf5 library is needed for linking any applications, so it's a natural dependency.

Tested:
VS2005 on WinXP
VNET on WinXP