Purpose: Backport of Windows project updates from trunk: Rename Windows HL Example projects to match other platforms
Description:
The collection of HL C examples that we distribute for Windows is incomplete.  I wasn't aware of it until Pedro pointed it out, because we use different (although similar) names for the generated files.  This checkin changes the naming to follow Linux, in preparation to add the rest of the HL examples to Windows.

Tested:
VS2005 on WinXP
VS.NET on WinXP

On a side-note, this change was very easy to port over to the 1.8 branch due to some convenient svn commands.  To bring the equivalent changes from one branch to another, issue the command:
svn merge {src-branch-dir} {dest-branch-dir} -r PREV:HEAD
assuming that the changes you want happened in the last checkin