Purpose: Fix typo in our configure CYGWIN patch
Description:
Previously, we created a patch for Cygwin to skip checking for the timezone variable.  We made this change in both the trunk and hdf5_1_8 branch.  However, in the branch version, we seemed to have used wrong quotes in our check ('uname' rather than `uname`).  This corrects the typo

Tested:
Cygwin on WinXP