Description:     Switch v1 B-tree nodes from using previous cache to use the new journaling cache code.  This is a major switch for the cache callbacks.
    Switched the metadata caching code from having a pointer to the file it's
in to receiving a pointer to the file, when needed.  This was necessary in
order to avoid crashing when two files IDs were open on the same underlying
file and one of those files was closed with cache entries using the file
pointers.  Also took out the check in the caching code for reading off the
end of the file, which didn't play nicely with the multi-file VFD.

    Switching the cache from having a pointer internally to requiring one
externally meant tweaking almost all the test code.  :-/

    Added correct exit codes to cache2 tests that didn't have them already,
so the 'make check' will stop when they fail.

    Use the path to the h5diff in this build in the tools/h5recover testing
script, since we can't guarantee a user has HDF5 already installed.

    Assorted minor tweaks to get everything to compile more cleanly and pass
all the tests (on my Mac :-).

Tested on:
    Mac OS X (10.5.2) w/C++
    (more testing coming up shortly, I just didn't have my "rsync testbed" set
up for this branch when I started making changes to the code)