Purpose: Disable ohdr_gentst project by default on Windows.
Description:
The ohdr_gentst project exists to dynamically generate the test files for ohdr test, which also exist in the repository.  The test is only run on other platforms when the configure flag --enable-build-all is specified.  Moreover, the source code ohdr_gentst.c is marked as _DO_NO_DISTRIBUTE_ in the MANIFEST.

As a result, there was a problem in the release tarball that the Windows projects would fail when trying to build this project.  To correct this, we no longer build this by default.  We will also need to release a FAQ/patch for the HDF5 1.8.0 release.

Tested:
VS2005 on WinXP