Purpose: bug fix.
Description:
Removed a few debug print statments that was left in by mistake.
Removed initialization for macro system INSTRUMENT to match how it
behaves before.

Tested platforms:
h5committested (kagiso, smirom and linew).  It failed in fortran
test in smirom but I believed that was not caused by my changes
in configure.in.