Purpose: Add hdf5.lib dependency to Windows C++ example project "compoundtest".
Description:
A user brought it to our attention that one of our C++ example projects was missing a linker dependency to the hdf5 library.  Somehow the project will still build without it using Visual Studio 6.0, but not VS2005.  This check-in simply adds the dependency to the "Release" configuration.

Tested:
VS6 on WinXP
VS2005 on WinXP