Purpose: Reset fill value in correct hyperslab selection (fixes bug 1155)
Description:
This bug was previously fixed, but the fixed was modified slightly from the 1.8 branch and the trunk.  The problem is that after we write a particular byte to manipulate the fill values, we need to write it back before reading and comparing to the expected value.  The previous change makes the second write after the hyperslab is manipulated, so the initial hyperslab is still incorrect.

Tested:
VS2005 on WinXP (32- and 64-bit)
h5committest