A variety of changes:
1) Removed references to H5AC2__CURR_JNL_CONFIG_VER from H5Fget_jnl_config()
   and H5Pget_jnl_config(), and also references to
   H5AC__CURR_CACHE_CONFIG_VERSION from H5Fget_mdc_config() and
   H5Pget_mdc_config().

2) Removed H5Pset_journal() from H5C2journal.c, and modified test
   code to use H5F/Pget/set_mdj_config() instead.

3) Implemented support for callbacks on metadata journaling status change
   along with the associated registration / deregistration calls and
   associated test code.

4) Fixed bug in journaling shutdown exposed by 3 above.

Tested and passed on Phoenix (serial), Linew (serial), and Kagiso (parallel).
However, while I was testing there were a couple of checkins, forcing 
an update and second round of testing.

On the second round, tested and passed on Phoenix (serial) and
Kagiso (parallel), but failed on Linew (serial).

As best I can tell, this was caused by Mike M's checkin -- which 
broke the smoke checks in cache2_journal on Linew but not Phoenix
or Kagiso.  A typical delta in the architype files follows:

linew.hdfgroup.uiuc.edu% diff -ctw cache2_journal_sc00_000.jnl tmp/cache2_journal_sc00_000.jnl
*** cache2_journal_sc00_000.jnl Fri Aug 22 08:28:49 2008
--- tmp/cache2_journal_sc00_000.jnl     Fri Aug 22 05:08:41 2008
***************
*** 1,5 ****
! 0 ver_num 1 target_file_name cache_journal_test.h5 creation_date Fri Aug 22 human_readable 1
! E eoa_value 0x0
  C comment Begin transaction on transaction 1.0.
  1 bgn_trans 1
  2 trans_num 1 length 1 base_addr 0x401 body  01
--- 1,5 ----
! 0 ver_num 1 target_file_name cache_journal_test.h5 creation_date Wed Aug 20 human_readable 1
! E eoa_value 0x772a9c01
  C comment Begin transaction on transaction 1.0.
  1 bgn_trans 1
  2 trans_num 1 length 1 base_addr 0x401 body  01

As you can see, it looks like garbage is getting into the first 
eoa write on Linew.

I'm checking in anyway, as Quincey needs my changes, and I will not 
have time to work on this for several days.

Mike:  Let me know if you are tackling this one -- if not, I'll deal with it.