Description:
Put the  DOUBLE PRECISION interfaces in a separate module. The interfaces are included/excluded depending on
the state of FORTRAN_DEFAULT_REALisDBLE_F which detects if the default REAL is DOUBLE PRECISION. This allows the library to be compiled with -r8 Fortran flag without the user needing to change the source code.