Bug 710: partial fix.
Temporary removed b2 from the --enable-debug=all list so that this option
will pass. "configure --help" lists the --enable-debug=all and users tend
to use it and end up with compile failures in the b2 debug code.  Since this
bug will not be fixed soon, this is a temporary fix to let the configure and
build work.

Tested:
h5committest plus "--enable-debu=all" test in kagiso.