Converted the fractal heap direct block cache client to use the revised cache.
Note that this conversion is not as efficient as it should be.  Specifically,
it does it more memcpy's between the metadata cache's on disk image of the 
direct block and the fractal heap's on disk image of the direct block than 
is absolutely essential.  Eventually, we will want to fix this -- probably
by allowing the metadata cache and the fractal heap direct block to share
a common on disk image of the direct block.  However, this will require 
extensions to the client / metadata cache interface, and some reworking of 
the fractal heap as well.

This checkin also includes Mike M's fix to the Linew specific bug mentioned
in my checkin of 22 Aug 2008.

Tested on Phoenix (serial debug and production), 
          Kagiso (parallel), and 
	  Linew (serial)