Purpose: Decrease size of points array in links test
Description:
In the external_set_elink_fapl2 links test, there is a large array of points which declared on the stack for testing.  Previously, the array was 1000 x 1000, which was too large for Visual Studio to handle.  As a result, we were getting stack overflows during the test.  We've reduced the number to 400, as this seems to be below the limit.  The exact number of points in the array isn't important to the test.

Tested:
VS2005 on WinXP