Bug 1529.
Description:
The MPE option is not working with the new mpich2. hdf5 configure is looking
for the routine CLOG_Init() in the library liblmpe.a. The old mpich has that
routine but the new mpich2 does not have it. I could not find any reference or
document of what the CLOG_Init() is or why it is needed.  The only log message
for adding the requirement for liblmpe.a was for the NCSA old AIX platform
that has retired and could not be used to verify the correctness.

Solution:
Remove the requirement of liblmpe.a.

Tested:
Just parallel with fortran on kagiso, jam and smirom.