Interrim checkin of shape same changes.
Modified H5S_select_shape_same() to accept as having the 
same shape some selections in spaces of different dimensionality.
Note that I reformated the function heavily in passing.  Must
return it to more standard format before the final checkin.

Also added test code associated with the change to test/tselect.c.
With the exception of removing some comments pointing out items
to be duscussed with Quincey, this code should be in close to its
final form.

Tested serial and parallel on Phoenix and parallel on Jam.

Serial test passes on Phoenix, but parallel test fails with a 
link error.

The link error does not appear in the parallel test on jam,
but a number of tool test failues do.