Bug fix: When gcc is used in the AIX platform, it failed to compile tconfig.c because it does not recognize the __int64 type.  The ibm-aix has the sizeof __int64=8 cached, as if all compilers supported this type.
Solution:
Removed the cached value for the sizeof __int64 so that during
configure, gcc can report it does not support it and not to use
it in the source code.

Tested:
NCSA Blue which is an AIX machine, using gcc or xlc compiler.
(The gcc still have failure but is being fixed as separated bugs.)