Added test code to verify that the expected writes, and only the expected writes take place in the parallel tests under the different metadata write strategies.
In essence, we do this by having each process call a "sync point done" 
callback at the end of the sync point, and pass it the list of entries 
that each process thinks should have been written at that sync point. 

That list is compared with the set of entries logged as written by the 
server process since the last time counters were reset, and an error 
is flagged if there is a mismatch.

Finally, the counters are reset, and the next epoch begins.

Note that this test adds a lot of overhead -- tests take roughly 
twice as long.  

Tested:  Phoenix: (amd64/debian) serial and parallel (with and without 
                                 trace file).

	Jam: serial and parallel

	Wren: serial and parallel (with and without trace file)

Note that at present, we only test transitions between metadata write 
strategies lightly and in passing.  This is probably OK, as shifts 
back and forth seem unlikely.  However, this point should be documented.