*** Post-code freeze change approved by Mike McGreevy ***
Removed linking to the BSD compatibility library (-lbsd-compat) on linux.
Linking to said library breaks the big test on CentOS 4.6 with gcc 4.6.2.
Although this is an unsual combination of OS and compiler, we're reverting
to the original behavior while we investigate the effect of the feature
macros.

Tested on jam and koala, incl. parallel and threadsafe.