Added code to H5Fget_file_image() to fail when invoked on the split/multi file drivers.
Added error rejection test code for H5Fget_file_image().

Note that we still have a problem with the family file driver and 
H5Fget_file_image() -- we will need to resolve this soon.

Added file image ops daisy chain parallel test.  Basic idea is to 
create a core file on process 0, pass its image to process 1 which 
validates it, modifies it, and passes the image to process 2, and 
so on until the modified image is passed back to process 0 and validated
yet again.  This should meet Mark Miller's parallel test requirement.

Tested serial on jam and both serial and parallel on Phoenix.  Didn't 
test parallel on jam as it seems that the parallel build environment
has been changed, so the file image ops branch no longer compiles and 
tests parallel on that machine.